* Refactor bug-related url construction
* Add captcha url (unused, yet)

* The second argument to `call/input-url' can be a one-argument function
  since the call doesn't send it any headers.

Still needs to be done: check the HTTP response for an error status, and
make it fail as a result (currently it just assumes that any response is
fine, which can change in case of a captcha failure).  Then, add a
captcha challenge if one is found, otherwise ignore.

original commit: 4276898a67a876282dba3c8beb1f110e55af82dd