Update insert-large-letters to use new GUI types
This doesn't change any functionality, it just uses
the more accurate GUI types and removes some type
annotations that are not necessary anymore.

It also removes the bitmap-message.rkt file because
the contents are now in insert-large-letters.rkt. I
think this class was kept separate only because TR
couldn't handle class definitions.