fix bug in rendering of bitmaps
When a bitmap is printed, the bytes are copied from the bitmap into an internal
table in order to avoid constructing them twice. the copy of bytes lived
too long, however, so when the same bitmap was converted those bytes
would be reused, even if the bitmap had been modified in the interim.

this commit evicts the bytes from the table as soon as they are used to
avoid this problem (this commit may also avoid leaks in some cases)

closes PR 15290