Display tooltips that are too large, and update docs on positions.
Previously, the example code in the documentation for tooltips did not
work unless there was additional whitespace at the end of the
module. This occurred for two reasons: tooltips outside the valid
range for editor positions were ignored, and tooltips are given in
editor positions, which differ from syntax object positions, so the
tooltip stuck out past the end of the file and was ignored.

This commit does two things to fix this:
 1. Tooltip regions are truncated, rather than ignored, when they
    exceed the file length.
 2. The docs mention the difference in indexing, and the sample code
    contains the necessary sub1 calls.

This solution is backwards-compatible - some non-working programs now
work, but working programs won't change the semantics of tooltip
placement.