fix "jump to next bound occurrence" (and related operations) so they ignore `require`-based arrows
This means you cannot step through the bindings coming from
a require anymore (which seems like a minor loss) but it also means
that programs like this one behave sanely:

  #lang racket/base
  (module+ test (require (submod "..")))
  (define state+ 11)
  (define add-tile 12)
  (module+ test state+ add-tile)

Before this change, c:x;n would go from the `state+` definition
to the `state+` reference to the `add-tile` reference because of
the arrows that start at the `(require (submod ".."))`. Now,
because those arrows are ignored, we don't get that behavior.