try to avoid concurrency problems in test
Some tests seem to be failing for scheduling reasons. It's not clear
why. The main strategy for avoiding concurrency issues is to queue a
low-priority callback on the grounds that anything a GUI needs to be
doing will be done by the time it gets around to a low-priority
callback. Maybe there's a `yield` somewhere that lets low-priority
callbacks through unexpectedly, or something like that. Using `(sync
(system-idle-evt))` is a lower-level way to make sure that nothing
else is going on, so maybe it will help.