Adjust drracket to use the unit-based errortrace-annotate
closes racket/htdp#103

This also changes how test coverage works and currently does not
try to cover module expressions (but just their bodies) because
of a mysterious tainting problem. The code that I believe is correct
is left commented out in a case in the test-coverage-point function

Also adds drracket:debug:make-debug-compile-handler export and other
minor bug fixes and improvements.