strengthen requirement of path in make-traversal
The current requirement of path in make-traversal is that it is
either #f or a path-string? However, because we parameterize
current-load-relative-directory to be this path, providing
any non complete-path? will result in a contract error.
This PR thus strengthens the requirement so that the provided path
must be complete-path?

The PR also fixes the show-content procedure so that
it satisfies the contract. Note that we need additional
parameterization of current-load-relative-directory so that
macro expansion can work correctly on relative path source like:

```
;; main.rkt
(require drracket/check-syntax)
(show-content "test/a.rkt")

;; test/a.rkt
(require "b.rkt")

;; test/b.rkt
1
```