improve the way drracket finds the relevant arrows when the start and end of the selection are the same
also improve the messages when someone tries to rename multiple
variables at once. It is dubious that this should even work the way it
does but in this program:

```racket
  #lang racket
  (define a 1)
  (define b 2)
  a b
```

if someone selects both `a` and `b` and then does a rename, at least
now they'll see that they're renaming two variables.

related to #414