use editor:silent-cancel-on-save-file-out-of-date? instead of editor:doing-autosave?
when automatically saving, we need to avoid asking the user questions
and the editor:doing-autosave? does that, but it also does other
things; in particular, it disables some of the autoload functionality
Specially, setting the editor:doing-autosave? parameter indicates that
the file being saved to isn't the primary file of this buffer, which
is wrong, it is still the primary save file in this situation. This
caused the autoload to get confused.