Fix error highlighting when background expansion is enabled
Currently, under the following conditions:

1. Errortrace is disabled
2. Background expansion is enabled, and it has finished expanding the program
3. The editor is unsaved (not named).

DrRacket will fail to highlight error. This is because the expanding
place uses a dummy srcloc "unsaved editor" which does not match the port
name of any editors, so the editors refuse to recognize the srcloc.

Because the port name of an editor is an uninterned symbol, passing it
to the expanding place doesn't look like a good idea. Instead, we
continue to use the dummy srcloc and adjust its source to be the port
name later.

Fixes #455

.