Improve interaction between submodules and require arrows.
Specifically, track whether each submodule has #f as the lang name not
as check syntax recurs and compare the module context at binders with
that at references and to avoid adding arrows from binders that are in
more-nested modules than the references. Also, some minor cleanups of the
code along the way.