generalize `drracket:define-popup` to callbacks
Predictably, having just a pair of symbol options is not configurable
enough for the popup "define" menu. Remove the option symbols and
instead support callback functions for the definition-finding and
name-extracting tasks. Each callback receives the default
implementation, in case it finds that helpful, and the former option
symbols have turned into optional keyword arguments for the default
definition-finding implementation.