Add a `continue` string to region style comment spec
This was discussed earlier but got left out.

This is relevant for a tool that allows "filling" comments (i.e.
adding/removing line breaks to fit a maximum column), as does Emacs,
where this is the value of the "comment-continue" variable. For line
comments this can obviously just be set to the same value as
comment-start, but for region comments it may differ (e.g. C++ might
use " *" and Scheme might use " ").