Add Resyntax Autofixer workflow
This workflow causes Resyntax to pick a random file in the repository once a week to analyze. Resyntax will create a pull request with its suggested fixes if the file has issues. Fixes are generated on Sunday at midnight, I think (cron schedules are not my strong suit). The frequency can be scaled up or down if desired. The workflow can also be triggered manually from the GitHub Actions tab of the repository. Pull requests created by Resyntax won't be merged automatically; human approval is still required for any changes to be made to the repository.