Revise lang-tools.scrbl to generalize from DrRacket
Mention a variety of tools.

When something applies to all tools, replace "DrRacket" with "tools";
otherwise retain the DrRacket specific information.

Change some contracts from "text%" or "racket:text<%>" to "or/c"
"color-textoid<%>" and explain why the latter is preferred to work
with more tools.

Re-order the initial overview itemlist to correspond with the
sections.

Some documentation had text outside the @language-info-def{} curly
parens; move it inside.

Auto-fill text to be consistent throughout.