Fix the Resyntax autofixer (#676)
The autofixer currently runs on a random file. However, if that file has nothing to fix, no changes are made so nothing happens. This pull request changes Resyntax to analyze the entire repository but uses the `--max-fixes` flag to limit the number of changes made to no more than ten fixes. Additionally, `xvfb-run` is removed since Resyntax no longer depends on the GUI framework.