add 3 units of padding around printed picts
A pict's bounding box does not bound the pixels of the pict's rendered
form. That kind of "bounding" box works well for composing pictures,
but it's confusing when lines and edges are cut off in a DrRacket
interaction. A pict's default conversion to bitmap or SVG form,
meanwhile, adds 3 units of padding around a pict to help avoid
confusion and edge effects, especially when rendering to Scribble
documents (where the passing is compensated by an inset return with
the bitmap to preserve alignment). Printing in DrRacket does not
convert a pict to a bitmap, but instead wraps it directly in a snip.

This commits adds the same padding around the pict within a snip for
its printed form. That padding extends the descent, so baseline
alignment still works. Snips don't support horizontal inset, at least
not in the same way as Scribble (or potentially other consumers of a
pict-to-bitmap conversion), but that seems less important in a
value-printing context.

With this change, the output of the following program looks nicer and
almost certainly is a better reflection on the programmer's intent:

```
(require pict)
(hline 10 0)
(rectangle 10 10)
```