create "drracket-core-lib" and "drracket-core" packages
The "drracket-core-lib" package is a minimal DrRacket implementation.
Since it lacks documentation, its dependencies are significantly fewer
than for the "drracket" package.

The "drracket-core" package adds documentation for DrRacket itself ---
which pulls in several other documents as dependencies.

The "drracket" package implies the "drracket-core" package and the
"drracket-core-lib" package. Much of the work in this commit reduces
the dependencies of the "drracket" package. In fact, the dependencies
for "drracket-core" are only slightly fewer. Still, the "drracket"
package keeps all features as before, even if only a small amount of
code is needed to enable some of thome, while "drracket-core" is free
to leave out old functionality (such as legacy languages, the ".plt"
installer, and separable functionality like the debugger).

DrRacket in "drracket-core-lib" still needs to adapt in small ways to
having no extra languages installed. Notably, the "Choose Language"
dialog should not list "Pretty Big" and "R5RS"when the supporting
packages are not included, and the "Add Teachpack..." menu item should
go away when there are no languages that support teachpacks.