sync "drracket-core" package version with "drracket"
Parts of "drracket" were moved to the "drracket-core" package. That
includes parts of "drracket" that have `version` annotations, but
moving then to "drracket-core" means the versions are mismatched.
Suncing the "drracket-core" version with "drracket" is not a perfect
solution, because the claim "added in version X of rdrracket-core" is
still not right, but it seems like the best available option right now.