improve the interactivity of check syntax's arrow drawing
The basic idea is to do the arrow drawing in a bitmap so that,
during a call to on-paint, we can just draw the bitmap. In addition,
incrementalize the drawing in the bitmap, carrying it out on multiple
event callbacks, taking care to give up on the drawing if some event
comes in that means there would be a different set of arrows

This seems to make check syntax's on-paint be pretty fast, usually
under 10 msec. From time to time, when there are a lot of arrows,
it can be slower to calculate the start and end positions of the
arrows (if there are a lot of them) and that's not incrementalized

Also, the tacked arrows and the untacked arrows have two separate bitmaps
so that, if there are a lot of tacked arrows, the drawing for them doesn't
need to be recomputed over and over